requirements.txt
flake8
pydocstyle
pyyaml
json
